refactor(materializer): extract scene-equivalence predicates (file #3)
Lift the pure conversation-patch equivalence helpers (areJsonLikeValuesEquivalent,
areSceneEntriesEquivalent, areActiveStreamingTailsEquivalent, areActivitiesEquivalent,
areActivitiesCompatibleForConversationPatch) into scene-equivalence.ts. Net 74/74, check 0.
